Typed HED code (#1424)
