+++ python/lib/physiological.py
@@ -1,8 +1,5 @@
 """This class performs database queries for BIDS physiological dataset (EEG, MEG...)"""
 
-from dataclasses import dataclass
-from functools import reduce
-
 from lib.database_lib.physiological_coord_system import PhysiologicalCoordSystem
 from lib.database_lib.point_3d import Point3DDB
 from lib.db.models.physio_file import DbPhysioFile
@@ -277,193 +274,3 @@ class Physiological:
         if blake2:
             # insert blake2b hash of task event file into physiological_parameter_file
             insert_physio_file_parameter(self.env, physiological_file, 'coordsystem_file_json_blake2b_hash', blake2)
-
-    @staticmethod
-    def get_additional_members_from_parenthesis_index(string_split, parentheses_to_find, end_index):
-        """
-        Helper method for determining AdditionalMembers for DB insert
-
-         :param string_split            : String array to search
-          :type string_split            : list[str]
-
-         :param parentheses_to_find     : Number of closing parentheses to find
-          :type parentheses_to_find     : int
-
-         :param end_index               : Current array index, to look back from
-          :type end_index               : int
-
-         :return                        : Number of additional members in group
-          :rtype                        : int
-
-         """
-        left_to_find = parentheses_to_find
-        sub_string_split = string_split[(len(string_split) - end_index - 1):]
-        additional_members = 0
-
-        for element_index, split_element in enumerate(sub_string_split):
-            left_to_find -= split_element.count(')')
-            left_to_find += split_element.count('(') if element_index > 0 else 0
-            if left_to_find == 1 and split_element.endswith(')'):
-                additional_members += 1
-            if left_to_find < 1:
-                return additional_members
-        return 0
-
-    @dataclass
-    class TagGroupMember:
-        hed_tag_id: int | None
-        has_pairing: bool
-        additional_members: int
-        tag_value: str | None = None
-
-        def __eq__(self, other):
-            return self.hed_tag_id == other.hed_tag_id and \
-                self.has_pairing == other.has_pairing and \
-                self.additional_members == other.additional_members
-
-    @staticmethod
-    def build_hed_tag_groups(hed_union, hed_string):
-        """
-        Assembles physiological event HED tags.
-
-        :param hed_union            : Union of HED schemas
-         :type hed_union            : any
-
-        :param hed_string           : HED string
-         :type hed_string           : str
-
-        :return                     : List of HEDTagID groups
-         :rtype                     : list[TagGroupMember]
-        """
-        # TODO: VALIDATE HED TAGS VIA SERVICE
-        # hedDict = utilities.assemble_hed_service(data_dir, event_tsv, event_json)
-
-        # NOT SUPPORTED: DEFS & VALUES
-
-        # TODO: TRANSACTION THAT ROLLS BACK IF HED_TAG_ID LIST MATCHES (CONSIDER ADDING ADDITIONAL
-        # + HP TO IT)
-
-        string_split = hed_string.split(',')
-        group_depth = 0
-        tag_groups = []
-        tag_group = []
-
-        for element_index, split_element in enumerate(string_split.__reversed__()):
-            additional_members = 0
-            if group_depth == 0:
-                if len(tag_group) > 0:
-                    tag_groups.append(tag_group)
-                    tag_group = []
-
-            element = split_element.strip()
-            right_stripped = element.rstrip(')')
-            left_stripped = right_stripped.lstrip('(')
-            num_opening_parentheses = len(right_stripped) - len(left_stripped)
-
-            has_pairing = element.startswith('(') and (
-                group_depth == 0 or not element.endswith(')')
-            )
-
-            if has_pairing:
-                additional_members = \
-                    Physiological.get_additional_members_from_parenthesis_index(string_split, 1, element_index)
-
-            hed_tag_id = Physiological.get_hed_tag_id_from_name(left_stripped, hed_union)
-            tag_group.append(Physiological.TagGroupMember(hed_tag_id, has_pairing, additional_members))
-
-            for i in range(
-                0 if group_depth > 0 and element.startswith('(') and element.endswith(')') else 1,
-                num_opening_parentheses
-            ):
-                has_pairing = True
-                additional_members = \
-                    Physiological.get_additional_members_from_parenthesis_index(string_split, i + 1, element_index)
-                tag_group.append(Physiological.TagGroupMember(None, has_pairing, additional_members))
-            group_depth += (len(element) - len(right_stripped))
-            group_depth -= num_opening_parentheses
-        if len(tag_group) > 0:
-            tag_groups.append(tag_group)
-
-        return tag_groups
-
-    @staticmethod
-    def standardize_row_columns(row):
-        """
-        Standardizes LORIS-recognized events.tsv columns to their DB column name
-
-       :param row                  : A row item from the events.tsv
-        :type row                  : dict
-
-       :return: Standardized row
-        :rtype: dict
-       """
-        standardized_row = {}
-        recognized_event_fields = [
-            'Onset', 'Duration', 'TrialType',
-            'ResponseTime', 'EventCode',
-            'EventSample', 'EventType'
-        ]
-        for column_name in row:
-            column_value = row[column_name]
-            if column_value is None:
-                continue
-
-            stripped_name = column_name.replace('_', '')
-            try:
-                field_index = list(map(lambda f: f.lower(), recognized_event_fields)).index(stripped_name)
-                column = recognized_event_fields[field_index]
-            except ValueError:
-                column = 'EventValue' if (column_name == 'value' or column_name == 'event_value') else column_name
-            standardized_row[column] = column_value
-
-        return standardized_row
-
-    @staticmethod
-    def filter_inherited_tags(row, tag_groups, dataset_tag_dict, file_tag_dict):
-        """
-        Filters for tags inherited from events.json
-
-        :param row                  : A row item from the events.tsv
-         :type row                  : dict
-        :param tag_groups           : Tag groups to filter
-         :type tag_groups           : list[list[TagGroupMember]
-        :param dataset_tag_dict     : Dict of dataset-inherited HED tags
-         :type dataset_tag_dict     : dict
-        :param file_tag_dict        : Dict of subject-inherited HED tags
-         :type file_tag_dict        : dict
-
-        :return: List of tag groups not inherited from events.json
-         :rtype: list[list[TagGroupMember]
-        """
-        # TODO: Overwrite dataset tags with file tags
-        # Only dataset tags currently supported until overwrite
-        standardized_row = Physiological.standardize_row_columns(row)
-        inherited_tag_groups = reduce(lambda a, b: a + b, [
-            dataset_tag_dict[column_name][standardized_row[column_name]]
-            for column_name in standardized_row
-            if column_name in dataset_tag_dict
-            and standardized_row[column_name] in dataset_tag_dict[column_name]
-        ], [])
-        return filter(
-            lambda tag_group: not any(
-                len(tag_group) == len(inherited_tag_group) and all(
-                    tag_group[i] == inherited_tag_group[i]
-                    for i in range(len(tag_group))
-                )
-                for inherited_tag_group in inherited_tag_groups
-            ),
-            tag_groups
-        )
-
-    @staticmethod
-    def get_hed_tag_id_from_name(tag_string, hed_union):
-        hed_tag_id = None
-        if tag_string is not None:
-            leaf_node = tag_string.split('/')[-1]  # LIMITED SUPPORT FOR NOW - NO VALUES OR DEFS
-            if len(tag_string) > 0:
-                hed_tag = next(filter(lambda tag: tag['Name'] == leaf_node, list(hed_union)), None)
-                if not hed_tag:
-                    print(f'ERROR: UNRECOGNIZED HED TAG: {tag_string}')
-                    raise
-                hed_tag_id = hed_tag['ID']
-        return hed_tag_id
