+++ python/lib/utilities.py
@@ -2,9 +2,7 @@
 
 import csv
 import filecmp
-import io
 import os
-import re
 import shutil
 import sys
 import tarfile
@@ -14,7 +12,6 @@ from datetime import datetime
 import loris_utils.crypto
 import mat73
 import numpy
-import requests
 import scipy.io
 from typing_extensions import deprecated
 
@@ -269,50 +266,3 @@ def remove_empty_folders(path_abs):
     for path, _, _ in walk[::-1]:
         if len(os.listdir(path)) == 0:
             os.rmdir(path)
-
-
-def assemble_hed_service(data_dir, event_tsv_path, event_json_path):
-    # Using HED Tool Rest Services to assemble the HED Tags
-    # https://hed-examples.readthedocs.io/en/latest/HedToolsOnline.html#hed-restful-services
-
-    # Request CSRF Token & session cookie
-    request_token_url = 'https://hedtools.ucsd.edu/hed/services'
-    token_response = requests.get(request_token_url)
-
-    cookie = token_response.headers['Set-Cookie']
-    token = re.search(r'csrf_token" value="(.+?)"', token_response.text).group(1)
-
-    # Define headers for assemble POST request, containing token and cookie
-    headers = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "X-CSRFToken": token,
-        "Cookie": cookie
-    }
-
-    # Read event files as str
-    event_json_text = open(data_dir + event_json_path).read()
-    event_tsv_text = open(data_dir + event_tsv_path).read()
-
-    # Define request parameters
-    params = {
-        'service': 'events_assemble',
-        'schema_version': '8.0.0',
-        'json_string': event_json_text,
-        'events_string': event_tsv_text,
-        'check_for_warnings': 'off',
-        'expand_defs': 'on',
-        'columns_included': ['onset']
-    }
-
-    # Make the request to assemble
-    request_assemble_url = 'https://hedtools.ucsd.edu/hed/services_submit'
-    assemble_response = requests.post(
-        request_assemble_url, headers=headers, json=params
-    )
-
-    # get assembled results as dictionary
-    data = assemble_response.json()['results']['data']
-    results = list(csv.DictReader(io.StringIO(data), delimiter='\t'))
-
-    return results
